Adding Functionality for Selecting and dragging
Not what they wanted @__@ Use mouse click to select the closest particle
to the cursor and use mouse drag to apply force on that particle. Let
the force vector be the vector between the position of the particle when
the mouse clicks and the position of the cursor when the mouse releases.